+++ NEWS.md
@@ -19,7 +19,8 @@ vignette: >
 - ee_Authenticate added.
 - Preventive message added to ee_get_date_ic (thanks @zackarno)
 - Critical bug solved in ee_check, see changes [here](https://github.com/r-spatial/rgee/pull/251/commits/9f6eeb00d5f26c48b7c95aa6c8a4d2feee04c795) (thanks @MartinHoldrege)
-
+- rgee use 'auth_mode = "notebook"' by default.
+- 'ee_clean_user_credentials' added. It will help users the remove their deprecated credentials. It remove the credentials rather than overwrite them as ee_Authenticate does.
 # rgee 1.1.3
 
 - Fixed a bug in ee_as_sf  (@MatthieuStigler).